requirements.txt
Django==1.8.4
djangorestframework==3.2.3

